Add installer targets to makefile and make modules local
Added movescripts, zipinstaller, installer targets to Makefile,
they are useful for making installer process automatic.
To create the installer run 'make installer' on deb distro.

Had to change modules import to local, prepended them with '.'
in main.py and calculator.py files.

Had to add main() function to main.py as an entry point for our app.

Added requirements.txt to src/. It's a good practice and might be needed.