Add installer folder with essential debian config files
Added installer folder to the main directory. In which are placed
essential config files for building debian pacakges.

setup.py is a general python script for building python modules.
control is the specification of a debian's package.
compat and rules more specify how the package is built.
changelog is needed for the build.

Build is made by using 'debuild' in source directory. Output files
are placed inside installer folder.